+++ weblate/settings_docker.py
@@ -1353,6 +1353,8 @@ ENABLE_SHARING = get_env_bool("WEBLATE_ENABLE_SHARING")
 
 EXTRA_HTML_HEAD = os.environ.get("WEBLATE_EXTRA_HTML_HEAD", "")
 
+UNUSED_ALERT_DAYS = get_env_int("WEBLATE_UNUSED_ALERT_DAYS", 365)
+
 # Wildcard loading
 for name in os.environ:
     if name.startswith("WEBLATE_RATELIMIT_") and name.endswith(