+++ docs/changes.rst
@@ -16,6 +16,7 @@ Not yet released.
 * Unchanged translation check can honor untranslatable terms from glossary.
 * Added automatic fixup for Devanagari danda.
 * Allow downloading project translation memory per language.
+* Added new alert on unused components.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/93?closed=1>`__.